Specific UpdateCoordinator for retrieving real time values
We can access each endpoint only 1 per minute.

The `getDevRealKpi` call has a `devTypeId` argument which is required, but this
makes that we need to do a call to the same endpoint per device type.

The DeviceRealKpiDataCoordinator will do the call per device type with a delay
of 1 minute in between each device type group.

This means that the more different device groups you have the less realtime the
data will be.

Related issues:
* https://github.com/tijsverkoyen/HomeAssistant-FusionSolar/issues/25
* https://github.com/tijsverkoyen/HomeAssistant-FusionSolar/issues/26