Use /thirdData/stations for https://au5.fusionsolar.huawei.com
It seems like getStationList is deprecated for certain devices.

I don't know if the check is any good. Probably it would be better to have some
sort of selector in the config flow that let the user decide if
`thirdData/stations` or `thirdData/getStationList` should be used...

This should fix https://github.com/tijsverkoyen/HomeAssistant-FusionSolar/issues/68