Redo call is getStationList is not available
Because Huawei Fusion Solar is not helping I just do a call to getStationList.
If it fails retry with stations.

It seems like getStationList is deprecated but they refuse to tell me how it
can be checked programatically. Therefore we need to do 2 calls. In a few
months the order should probably be switched. So getStationList is the fallback